@@ -7,19 +7,19 @@
   },
   "author": "Vinay Pulim <v@pulim.com>",
   "dependencies": {
-    "axios-ntlm": "^1.2.0",
-    "debug": "^4.3.2",
-    "formidable": "^3.2.4",
+    "axios-ntlm": "^1.4.2",
+    "debug": "^4.3.4",
+    "formidable": "^3.5.1",
     "get-stream": "^6.0.1",
     "lodash": "^4.17.21",
-    "sax": ">=0.6",
+    "sax": "^1.3.0",
     "strip-bom": "^3.0.0",
-    "uuid": "^8.3.2",
+    "uuid": "^9.0.1",
     "whatwg-mimetype": "3.0.0",
-    "xml-crypto": "^3.0.0"
+    "xml-crypto": "^5.1.1"
   },
   "peerDependencies": {
-    "axios": "^0.27.2"
+    "axios": "^1.6.5"
   },
   "repository": {
     "type": "git",
@@ -41,7 +41,7 @@
     "cover": "nyc --extension=.ts --reporter=lcov --reporter=html --reporter=text mocha --timeout 15000 --exit test/*-test.js test/security/*.js",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js -v",
     "docs": "typedoc --out docs",
-    "test": "mocha --timeout 15000 --bail --exit test/*-test.js test/security/*.js"
+    "test": "mocha --timeout 15000 --bail --exit test/client-test.js test/security/*.js"
   },
   "keywords": [
     "soap"
@@ -49,36 +49,36 @@
   "license": "MIT",
   "devDependencies": {
     "@types/axios": "^0.14.0",
-    "@types/debug": "^4.1.7",
-    "@types/express": "^4.17.13",
-    "@types/formidable": "^2.0.4",
-    "@types/lodash": "^4.14.172",
+    "@types/debug": "^4.1.12",
+    "@types/express": "^4.17.21",
+    "@types/formidable": "^3.4.5",
+    "@types/lodash": "^4.14.202",
     "@types/node": "^11.15.54",
-    "@types/request": "^2.48.7",
-    "@types/sax": "^1.2.3",
-    "@types/uuid": "^8.3.1",
-    "@types/whatwg-mimetype": "^2.1.1",
-    "body-parser": "^1.15.2",
-    "colors": "^1.3.3",
+    "@types/request": "^2.48.12",
+    "@types/sax": "^1.2.7",
+    "@types/uuid": "^9.0.7",
+    "@types/whatwg-mimetype": "^3.0.2",
+    "body-parser": "^1.20.2",
+    "colors": "^1.4.0",
     "coveralls": "^3.1.1",
-    "diff": "^4.0.1",
-    "doctoc": "^2.0.1",
-    "duplexer": "~0.1.1",
-    "express": "^4.16.4",
-    "finalhandler": "^1.1.1",
-    "glob": "^7.1.7",
-    "jshint": "^2.13.1",
-    "mocha": "^6.1.4",
-    "nyc": "^14.1.1",
-    "readable-stream": "~2.0.2",
-    "semver": "^5.6.0",
-    "serve-static": "^1.14.1",
+    "diff": "^5.1.0",
+    "doctoc": "^2.2.1",
+    "duplexer": "~0.1.2",
+    "express": "^4.18.2",
+    "finalhandler": "^1.2.0",
+    "glob": "^10.3.10",
+    "jshint": "^2.13.6",
+    "mocha": "^10.2.0",
+    "nyc": "^15.1.0",
+    "readable-stream": "~4.5.2",
+    "semver": "^7.5.4",
+    "serve-static": "^1.15.0",
     "should": "^13.2.3",
-    "sinon": "^1.17.7",
-    "source-map-support": "^0.5.10",
-    "timekeeper": "^2.1.2",
-    "tslint": "^5.18.0",
-    "typedoc": "^0.23.10",
-    "typescript": "^4.7.4"
+    "sinon": "^17.0.1",
+    "source-map-support": "^0.5.21",
+    "timekeeper": "^2.3.1",
+    "tslint": "^6.1.3",
+    "typedoc": "^0.25.7",
+    "typescript": "^4.9.5"
   }
 }
